+++ tests/test_checker/test_noqa.py
@@ -179,6 +179,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS445': 1,
     'WPS446': 1,
     'WPS447': 1,
+    'WPS448': 1,
 
     'WPS500': 1,
     'WPS501': 1,