+++ wemake_python_styleguide/violations/best_practices.py
@@ -67,6 +67,7 @@ Summary
    WrongNamedKeywordViolation
    ApproximateConstantViolation
    StringConstantRedefinedViolation
+   IncorrectExceptOrderViolation
 
 Best practices
 --------------
@@ -119,6 +120,7 @@ Best practices
 .. autoclass:: WrongNamedKeywordViolation
 .. autoclass:: ApproximateConstantViolation
 .. autoclass:: StringConstantRedefinedViolation
+.. autoclass:: IncorrectExceptOrderViolation
 
 """
 
@@ -1857,3 +1859,46 @@ class StringConstantRedefinedViolation(ASTViolation):
 
     error_template = 'Found alphabet as strings: {0}'
     code = 447
+
+
+@final
+class IncorrectExceptOrderViolation(ASTViolation):
+    """
+    Forbids the use incorrect order of ``except``.
+
+    Note, we only check for built-in exceptions.
+
+    Reasoning:
+        Using incorrect order of exceptions is error-prone, since
+        you end up with some unreachable exception clauses.
+
+    Solution:
+        Use correct order of exceptions.
+
+    Example::
+
+        # Correct:
+        try:
+            ...
+        except ValueError:
+            ...
+        except Exception:
+            ...
+
+        # Wrong:
+        try:
+            ...
+        except Exception:
+            ...
+        except ValueError:
+            ...
+
+    See also:
+        https://bit.ly/36MHlzw
+
+    .. versionadded:: 0.13.0
+
+    """
+
+    error_template = 'Found incorrect exception order'
+    code = 448